[qt] Break cross-plugin symbol dependencies (plugin isolation phase 1)
Two Category 2 violations fixed:

1. ores.qt.trading → ores.qt.party:
   Add IBusinessUnitBrowser abstract interface to ores.qt.api. Replace
   BusinessUnitController* with IBusinessUnitBrowser* in
   OrgExplorerMdiWindow. BusinessUnitController implements the interface.
   Remove ores.qt.party.lib from ores.qt.trading's PUBLIC link deps.

2. ores.qt.data_transfer → ores.qt.refdata:
   Move the 8 DQ catalogue controllers (DataDomain, SubjectArea, Catalog,
   DatasetBundle, Methodology, OriginDimension, NatureDimension,
   TreatmentDimension) from DataTransferPlugin into RefdataPlugin where
   they logically belong. RefdataPlugin::setup_menus() now builds the
   Data Catalogue submenu on smc.data_transfer_menu. DataTransferPlugin
   becomes a thin shell that only owns the menu handle. Remove
   ores.qt.refdata.lib from ores.qt.data_transfer's link deps entirely.
