+++ projects/ores.qt.data_transfer/src/CMakeLists.txt
@@ -47,16 +47,11 @@ set_target_properties(${lib_target_name} PROPERTIES
 target_include_directories(${lib_target_name} PUBLIC
     ${ORES_QT_DATA_TRANSFER_DIR}/include)
 target_include_directories(${lib_target_name} PRIVATE
-    "${CMAKE_CURRENT_BINARY_DIR}/${lib_target_name}_autogen/include"
-    "${CMAKE_SOURCE_DIR}/projects/ores.qt.refdata/include"
-    "${CMAKE_SOURCE_DIR}/projects/ores.qt/include")
+    "${CMAKE_CURRENT_BINARY_DIR}/${lib_target_name}_autogen/include")
 
 target_link_libraries(${lib_target_name}
     PUBLIC
-        ores.qt.api.lib
-        ores.qt.refdata.lib
-    PRIVATE
-        ores.qt.lib)
+        ores.qt.api.lib)
 
 install(TARGETS ${lib_target_name}
     LIBRARY DESTINATION bin
